[refactor][yunnij] - refactor kakao ConfigurationProperties
