+++ src/main/java/com/offnal/shifterz/global/config/JwtConfig.java
@@ -1,5 +1,6 @@
-package com.offnal.shifterz.jwt;
+package com.offnal.shifterz.global.config;
 
+import com.offnal.shifterz.jwt.JwtProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
