+++ src/main/java/com/offnal/shifterz/kakao/KakaoProperties.java
@@ -1,0 +1,9 @@
+package com.offnal.shifterz.kakao;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "kakao")
+public record KakaoProperties(
+        String clientId,
+        String redirectUri
+) {
+}
